Splits graph CUnit tests into separate files
