+++ Source/graph/CUnit/CU_graph_bfs.c
@@ -1,75 +1,10 @@
-//Zero Hanami (C) 2021: CU_graph.c
+//Zero Hanami (C) 2021: CU_graph_bfs.c
 
 #include <stdio.h>
 #include "CU_graph.h"
 #include "0hana/random/random.h"
 #define VERTEX_LIMIT 100
 #define EDGE_LIMIT -1
-
-int init_graph_suite(void) { return 0; }
-int clean_graph_suite(void) { return 0; }
-
-void CU_graph_random(void) {
-
-	unsigned int const Failures = CU_get_number_of_failures();
-
-	for(size_t I = 0; I < 1000; I++) {
-
-		struct graph * const G = graph_random(VERTEX_LIMIT, EDGE_LIMIT);
-		CU_ASSERT_PTR_NOT_NULL_FATAL(G);
-		CU_ASSERT(G->Terms < VERTEX_LIMIT);
-
-		for(size_t V = 0; V < G->Terms; V++) {
-			CU_ASSERT(G->Term[V].Links < G->Terms);
-
-			if(G->Term[V].Links > 0) CU_ASSERT(G->Term[V].Link[0].Term < G->Terms);
-			for(size_t E = 1; E < G->Term[V].Links; E++) {
-				CU_ASSERT(G->Term[V].Link[E].Term < G->Terms);
-				CU_ASSERT(G->Term[V].Link[E].Term > G->Term[V].Link[E - 1].Term);
-			}
-		}
-		
-		graph_free(G);
-
-		CU_ASSERT_EQUAL_FATAL(Failures, CU_get_number_of_failures());
-	}
-}
-
-void CU_graph_transpose(void) {
-
-	unsigned int const Failures = CU_get_number_of_failures();
-
-	for(size_t I = 0; I < 1000; I++) {
-
-		struct graph * const G = graph_random(VERTEX_LIMIT, EDGE_LIMIT);
-		CU_ASSERT_PTR_NOT_NULL_FATAL(G);
-
-		struct graph * const T = graph_transpose(G);
-		if(!T) {
-			graph_free(G);
-			CU_FAIL_FATAL("Failed to allocate Transpose graph -- stopping test");
-		}
-
-		CU_ASSERT_EQUAL(T->Terms, G->Terms);
-
-		//Test that G edge destinations as T edge sources point to the vertex of that edge in G
-		size_t T_Term_Link_Iterator[G->Terms];
-		for(size_t V = 0; V < G->Terms; V++) T_Term_Link_Iterator[V] = 0;
-		for(size_t V = 0; V < G->Terms; V++) {
-			for(size_t E = 0; E < G->Term[V].Links; E++) {
-				#define G_DEST G->Term[V].Link[E].Term
-				CU_ASSERT_EQUAL(V, T->Term[G_DEST].Link[T_Term_Link_Iterator[G_DEST]++].Term);
-				#undef G_DEST
-			}
-		}
-
-		graph_free(T);
-		graph_free(G);
-
-		//Problem?:
-		CU_ASSERT_EQUAL_FATAL(Failures, CU_get_number_of_failures());
-	}
-}
 
 void CU_graph_bfs(void) {
 	/*
