+++ Source/graph/CUnit/CU_graph_suite.c
@@ -1,0 +1,10 @@
+//Zero Hanami (C) 2021: CU_graph.c
+
+#include <stdio.h>
+#include "CU_graph.h"
+#include "0hana/random/random.h"
+#define VERTEX_LIMIT 100
+#define EDGE_LIMIT -1
+
+int init_graph_suite(void) { return 0; }
+int clean_graph_suite(void) { return 0; }
